app/BE/requirements.txt
pymongo==4.11.1
python-dotenv==1.0.1
motor==3.7.0
uvicorn==0.34.0
fastapi==0.115.8
python-multipart==0.0.20
pydantic-settings==2.7.1
itsdangerous==2.2.0